--- src/main/kotlin/eu/ibagroup/zdevops/model/ResolvedZOSConnection.kt
@@ -1,14 +1,0 @@
-package eu.ibagroup.zdevops.model
-
-import java.io.Serializable
-
-data class ResolvedZOSConnection(
-  val name: String,
-  val url: String,
-  val username: String,
-  val password: String
-) : Serializable {
-  override fun toString(): String {
-    return "ResolvedZOSConnection(name=$name,url=$url)"
-  }
-}
