+++ src/main/kotlin/org/zowe/zdevops/classic/steps/SubmitJobStep.kt
@@ -1,0 +1,61 @@
+package org.zowe.zdevops.classic.steps
+
+import org.zowe.kotlinsdk.zowe.client.sdk.zosjobs.SubmitJobs
+import org.zowe.zdevops.Messages
+import org.zowe.zdevops.classic.AbstractBuildStep
+import org.zowe.zdevops.config.ZOSConnectionList
+import hudson.AbortException
+import hudson.Extension
+import hudson.Launcher
+import hudson.model.AbstractBuild
+import hudson.model.BuildListener
+import hudson.util.ListBoxModel
+import jenkins.model.GlobalConfiguration
+import org.kohsuke.stapler.DataBoundConstructor
+import java.io.PrintWriter
+import java.io.StringWriter
+
+class SubmitJobStep
+@DataBoundConstructor
+constructor(
+  connectionName: String,
+  val jobName: String
+) : AbstractBuildStep(connectionName) {
+  override fun perform(
+      build: AbstractBuild<*, *>,
+      launcher: Launcher,
+      listener: BuildListener,
+      zosConnection: org.zowe.kotlinsdk.zowe.client.sdk.core.ZOSConnection
+  ) {
+    runCatching {
+      listener.logger.println(Messages.zdevops_classic_ZOSJobs_submitting(jobName, zosConnection.host, zosConnection.zosmfPort))
+      val submitJobRsp = SubmitJobs(zosConnection).submitJob(jobName)
+      listener.logger.println(
+          Messages.zdevops_classic_ZOSJobs_submitted_success(
+              submitJobRsp.jobid,
+              submitJobRsp.jobname,
+              submitJobRsp.owner
+          )
+      )
+    }.onFailure {
+      val sw = StringWriter()
+      it.printStackTrace(PrintWriter(sw))
+      listener.logger.println(sw.toString())
+      throw AbortException(Messages.zdevops_classic_ZOSJobs_submitted_fail(jobName))
+    }
+  }
+
+
+  @Extension
+  class DescriptorImpl : Companion.DefaultBuildDescriptor(Messages.zdevops_classic_submitJobStep_display_name()) {
+    fun doFillConnectionNameItems(): ListBoxModel {
+      val result = ListBoxModel()
+
+      GlobalConfiguration.all().get(ZOSConnectionList::class.java)?.connections?.forEach {
+        result.add("${it.name} - (${it.url})", it.name)
+      }
+
+      return result
+    }
+  }
+}
