+++ src/main/kotlin/org/zowe/zdevops/model/ResolvedZOSConnection.kt
@@ -1,0 +1,14 @@
+package org.zowe.zdevops.model
+
+import java.io.Serializable
+
+data class ResolvedZOSConnection(
+  val name: String,
+  val url: String,
+  val username: String,
+  val password: String
+) : Serializable {
+  override fun toString(): String {
+    return "ResolvedZOSConnection(name=$name,url=$url)"
+  }
+}
